py_bot_zomboid/pyproject.toml
[tool.poetry]
name = "py-bot-zomboid"
version = "0.1.0"
description = ""
authors = ["Matthew Lewis <mlewis@setonhill.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
